Change more task_dist fields to 32-bits
The task_dist field was expanded from 16-bits to 32-bits in commit
a4473cf950ca2fbf072d80331edb7e3c3dd62b33
However the field size was not expanded in all places and translation
of old format data to the new format was not being done in many
places. This commit, along with commit
456911254c4701e6cb85cbbfe6bcf12edb39587b
addresses those shortcomings.

+++ slurm/slurmdb.h
@@ -799,7 +799,7 @@ typedef struct {
 	uint32_t suspended;
 	uint32_t sys_cpu_sec;
 	uint32_t sys_cpu_usec;
-	uint16_t task_dist;
+	uint32_t task_dist;
 	uint32_t tot_cpu_sec;
 	uint32_t tot_cpu_usec;
 	char *tres_alloc_str;
