+++ src/common/slurm_step_layout.h
@@ -68,7 +68,7 @@ extern slurm_step_layout_t *slurm_step_layout_create(const char *tlist,
 						     uint32_t node_cnt,
 						     uint32_t task_cnt,
 						     uint16_t cpus_per_task,
-						     uint16_t task_dist,
+						     uint32_t task_dist,
 						     uint16_t plane_size);
 
 /*
